Support graph modification via. exception throw
In order to support tasks being able to modify the graph an exception
can be thrown which is caught by the walker and causes the walker to
re-run using the provided replacement graph.

Use cases for this:

- loading tasks from a CI configuration which is generated in a previous
  task.
- support iterating over environment variables

An exception is thrown as it is an event which interupts the flow of the
graph and requires that the graph walker be reset, an alternative would
be to introduce a response and supply the modified graph in the response
from the task handler (this can still be an option later on).