Added interface method to task handler
Use an explicit method rather than depending on __invoke. The advantage
is that it's explicit, the disadvantage is that the end-user must
explicitly assert the type of the task that comes in, as PHP does not
currently support co-variance. But this also opens up the door to
adding a "rollback" method to handlers.