Move success/failure states outside of the state machine
It is very difficult to handle success / failure states in conjunction
with rescheduled tasks.

The node state is primarily used for scheduling (WAITING is ready to go,
BUSY is in progress, DONE/CANCELLED are finished, SCHEDULED is ready to
be scheduled.

The DONE (success) / FAILED states do not indicate if the node should be
rescheduled or not but both indicate that the node has finished
processing, success or not.

FAILED nodes will now have the state DONE, but there is a new state, the
taskResult state, which has PENDING, SUCCESS, or FAILED